Capabilities/requirements.txt
boto3
chalice
requests
google-cloud-vision
pillow
pytesseract